fix: avatar component test case failing
