+++ src/components/avatar/avatar.styles.ts
@@ -1,8 +1,7 @@
+import { AvatarSize } from 'boilerplate-react-native/src/types';
 import { useTheme } from 'native-base';
 import { StyleSheet } from 'react-native';
 import type { ViewStyle, ImageStyle, TextStyle } from 'react-native';
-
-import { AvatarSize } from './avatar';
 
 const sizeMap = {
   [AvatarSize.small]: 32,
