+++ src/types/avatar.ts
@@ -1,0 +1,9 @@
+export enum AvatarSize {
+  small = 'small',
+  medium = 'medium',
+  large = 'large',
+}
+export enum AvatarShape {
+  circle = 'circle',
+  square = 'square',
+}
