+++ src/types/index.ts
@@ -4,3 +4,4 @@ export * from './async';
 export * from './auth';
 export * from './common';
 export * from './task';
+export * from './avatar';
